pyproject.toml
[tool.poetry]
name = "cqbot"
version = "0.3.0"
description = "just a chatting bot"
authors = ["Lycreal"]
license = "MIT"

readme = 'README.md'  # Markdown files are supported

repository = "https://github.com/Lycreal/cqbot"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0b5"
nonebot-adapter-onebot = "^2.1.1"
nonebot-plugin-manager = "^0.5.9"
nonebot-plugin-help = "^0.3.1"
nonebot-plugin-apscheduler = "^0.1.4"
pydantic = "^1.8.1"
lxml = "^4.9.1"
httpx = "^0.21.3"
opencv-python-headless = "^4.5.2"
numpy = "^1.22.0"
starlette = "^0.19.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-cov = "^2.12.0"
pytest-asyncio = "^0.16.0"
requests = "^2.25.1"
mypy = "^0.910"
nonebot-plugin-test = "^0.2.0"
nonebug = "^0.2.1"

[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/reference.html#configuration-options
minversion = "6.0"
testpaths = [
    "tests"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
